requirements.txt
spotipy
python-dotenv
smolagents
litellm
pyyaml
SpeechRecognition
PyAudio
openai-whisper
soundfile
edge-tts
pygame
rapidfuzz
openwakeword
pyaudio
numpy
ddgs